Extract ImagePanel class to use in image preprocessing examples
